refactor: make kernel db dumper a separate function from the command line utility
